+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/projects/ProjectsRendererI.java
@@ -54,6 +54,7 @@ import static net.splitcells.website.server.project.LayoutUtils.extendPerspectiv
 import static net.splitcells.website.server.projects.extension.GlobalChangelogExtension.globalChangelogExtension;
 import static net.splitcells.website.server.projects.extension.LayoutExtension.layoutExtension;
 import static net.splitcells.website.server.projects.extension.LayoutTreeExtension.layoutTreeExtension;
+import static net.splitcells.website.server.projects.extension.TestExtension.testExtension;
 import static net.splitcells.website.server.projects.extension.status.NetworkStatusRenderExtension.networkStatusRenderExtension;
 import static net.splitcells.website.server.projects.extension.ProjectsRendererExtensionMerger.projectsRendererExtensionMerger;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -162,7 +163,8 @@ public class ProjectsRendererI implements ProjectsRenderer {
             .withRegisteredExtension(globalChangelogExtension())
             .withRegisteredExtension(networkStatusRenderExtension())
             .withRegisteredExtension(layoutExtension())
-            .withRegisteredExtension(layoutTreeExtension());
+            .withRegisteredExtension(layoutTreeExtension())
+            .withRegisteredExtension(testExtension());
 
     private ProjectsRendererI(String name
             , ProjectRenderer fallbackRenderer
@@ -178,7 +180,7 @@ public class ProjectsRendererI implements ProjectsRenderer {
         final var normalizedSlash = path.replaceAll("//+", "/");
         final String rootMatched;
         if (config.possibleRootIndex().contains(normalizedSlash)) {
-            return normalizedSlash;
+            return config.rootIndex();
         } else {
             rootMatched = normalizedSlash;
         }